Disambiguate Revit 2026 ElementId overloads across MCP tools
Same fix as 99eef91 (create_dimension) and d5fe13e (create_room) applied
to every remaining DB.ElementId(int(...)) call site in the repo:

  revit_mcp/element_management.py
    - L114 delete_elements id resolution
    - L191 modify_element id resolution
    - L216 modify_element ElementId-storage parameter setter

  revit_mcp/element_creation.py
    - L84 _set_instance_parameters ElementId-storage setter

On Revit 2026 IronPython, DB.ElementId now has three overloads:
ElementId(BuiltInParameter), ElementId(BuiltInCategory), ElementId(Int64).
A bare Python int can box to any of them, so the call raises:

    TypeError: Multiple targets could match:
      ElementId(BuiltInParameter), ElementId(BuiltInCategory),
      ElementId(Int64)

Casting through System.Int64 nails down the Int64 overload.

These were latent until any caller actually exercised them with a real
integer id. element_management.delete_elements and modify_element are
high-traffic entry points — the bug would have surfaced on the next
real user call.

Also updates the example module in LLM.txt and README.md so anyone
copy-pasting starts with the correct pattern, with a comment explaining
why the Int64 cast is needed.

+++ revit_mcp/element_creation.py
@@ -9,6 +9,7 @@ Internally converted to Revit's decimal-feet via UnitUtils.
 
 from pyrevit import routes, revit, DB
 from utils import normalize_string, element_id_value
+from System import Int64
 import json
 import traceback
 import logging
@@ -81,7 +82,7 @@ def _set_instance_parameters(element, properties):
                 # Caller is responsible for unit conversion; accept raw float
                 p.Set(float(value))
             elif st == DB.StorageType.ElementId:
-                p.Set(DB.ElementId(int(value)))
+                p.Set(DB.ElementId(Int64(int(value))))
             written.append(name)
         except Exception as pe:
             logger.warning("Could not set parameter {}: {}".format(name, str(pe)))
